requirements.txt
cffi==1.15.0
Pillow==8.4.0
pillow-heif==0.1.4
pycparser==2.21
